// studio/datart-versions/datart-1.0.0-rc.1/frontend/src/app/types/ChartConfig.ts
export type FilterCondition = {
  name: string;
  type: FilterConditionType;
  value?:
    | Lowercase<keyof typeof FilterRelationType>
    | string
    | number
    | [number, number]
    | string[]
    | Array<RelationFilterValue>
    | TimeFilterConditionValue;
  visualType: string;
  operator?:
    | string
    | Lowercase<keyof typeof FilterRelationType>
    | Uncapitalize<keyof typeof FilterSqlOperator>;
  children?: FilterCondition[];
};
// ...
export type ChartDataSectionField = {
  uid?: string;
  colName: string;
  desc?: string;
  type: DataViewFieldType;
  category: Uncapitalize<keyof typeof ChartDataViewFieldCategory>;
  expression?: string;
  field?: string;
  sort?: SortFieldAction;
  alias?: AliasFieldAction;
  format?: FormatFieldAction;
  aggregate?: AggregateFieldActionType;
  filter?: FilterFieldAction;
  color?: ColorFieldAction;
  size?: number;
  path?: string[];
};
// ...
export type ChartConfigBase = {
  label?: string;
  key: string;
};

export type ChartDataConfig = ChartConfigBase & {
  type?: Lowercase<keyof typeof ChartDataSectionType>;
  allowSameField?: boolean;
  required?: boolean;
  rows?: ChartDataSectionField[];
  actions?: Array<ValueOf<typeof ChartDataSectionFieldActionType>> | object;
  limit?: null | number | string | number[] | string[];
  disableAggregate?: boolean;
  drillable?: boolean;
  drillContextMenuVisible?: boolean;
  options?: {
    [key in ValueOf<typeof ChartDataSectionFieldActionType>]: {
      backendSort?: boolean;
    };
  };
  replacedConfig?: ChartDataSectionField;
  // NOTE: keep field's filter relation for filter arrangement feature
  fieldRelation?: FilterCondition;
  // Runtime filters
  [RUNTIME_FILTER_KEY]?: ChartDataRequestFilter[];
};

export type ChartStyleSectionTemplate = {
  template?: ChartStyleSectionRow;
};

export type ChartStyleConfig = ChartConfigBase &
  ChartStyleSectionGroup &
  ChartStyleSectionTemplate;

export type ChartStyleSectionGroup = ChartStyleSectionRow & {
  rows?: ChartStyleSectionGroup[];
};
// ...
export type ChartI18NSectionConfig = {
  lang: string;
  translation: object;
};

export type ChartConfig = {
  datas?: ChartDataConfig[];
  styles?: ChartStyleConfig[];
  settings?: ChartStyleConfig[];
  interactions?: ChartStyleConfig[];
  i18ns?: ChartI18NSectionConfig[];
};